Updated functions.html up to 1.54f and logic to parse it
This new version of the file is now also included in the ImageJ repo and served in the imagej.net web. Functions up to 1.54f are included.
The logic that parses new functions families (`Color.`, `Image.`, `Property.` and `RoiManager.`) as well as the one that resolves relative links to full URLs was updated to match the web version.

See this discussion:
https://github.com/imagej/imagej-legacy/pull/301#issuecomment-1659304843